Fixed boss installation for newer boss versions (#117)
Boss 7.0.9.p01 had problems compiling analysis code because the workarea was not in CMTPATH, indicating an issue with the boss installation.
Investigation showed that the original content of a line in cmthome/requirements that is replaced by the install scripts changed, so the sed command making the replacement didn't recognize it anymore. Consequently, the line was not replaced, and the workara was unknown for those versions.
I adjusted the install script to be more inclusive, it should now recognize any (valid) version of that line and plop in the correct workarea properly.