Implemented automatic copying of TestRelease (#139)
As mentioned in issue https://github.com/j-reher/bossdocker/issues/135, test TestRelease missing in an empty WorkArea can be quite inconvenient.
Now, the container will check if the correct version of TestRelease for its BOSS version is present in the workarea, and if not, it is copied in there. This happens before even mounting the cvmfs directories, as part of dockerinit.sh.